+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 JS3 Changelog
 ===
 
-##### 0.6.x - &beta; (Future)
+##### 0.7.x - &beta; (Future)
 
 * Issue fixes
 * Property specific extensions
 * CSS to JS3 porting for an easy switch to JS3 for existing work
 * Sample JS3 files (ports of few standard CSS files)
 
-##### 0.5.x - &beta; (Next)
+##### 0.6.x - &beta; (Next)
 
+* Issue fixes
 * Full-fledged running example (Basic)
+
+##### 0.5.5 - &beta; (07 Mar 2014)
+
+* Issues fixed
+* Added support for declaration level debugging
+* Added support for interactive reset for dynamic style and style rules
+* Completed guides documentation
 
 ##### 0.4.9 - &beta; (27 Feb 2014)
 * Minor API standardization